# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Project Management Federation API
  version: 0.1.0
  description: |
    Backend REST, ActivityPub, C2S, S2S inbox, WebFinger, and delivery administration contract.

    Bounded REST list operations return JSON arrays for backward compatibility and expose a uniform
    offset-pagination contract through `Link`, `X-Pagination-Limit`, `X-Pagination-Offset`,
    `X-Pagination-Count`, and `X-Pagination-Has-More` response headers. The `Link` header preserves
    active filters and supplies RFC 8288 `prev` and `next` relations when applicable.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/me/notifications:
    get:
      tags: [Notifications]
      operationId: listCurrentUserNotifications
      security: [{bearerAuth: []}, {cookieAuth: []}]
      parameters:
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 50}}
        - {name: offset, in: query, schema: {type: integer, minimum: 0, default: 0}}
        - {name: unread, in: query, schema: {type: boolean, default: false}}
      responses:
        "200":
          description: Current user notifications.
          content:
            application/json:
              schema: {type: array, items: {$ref: "#/components/schemas/Notification"}}
        "400": {$ref: "#/components/responses/Error"}
  /api/v1/me/notifications/events:
    get:
      tags: [Notifications]
      operationId: streamCurrentUserNotifications
      security: [{bearerAuth: []}, {cookieAuth: []}]
      responses:
        "200":
          description: Server-sent notification event stream.
          content:
            text/event-stream:
              schema: {type: string}
  /api/v1/me/notifications/{id}/read:
    patch:
      tags: [Notifications]
      operationId: markNotificationRead
      security: [{bearerAuth: []}, {cookieAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200": {description: Updated notification., content: {application/json: {schema: {$ref: "#/components/schemas/Notification"}}}}
        "404": {$ref: "#/components/responses/Error"}
  /api/v1/me/notifications/read-all:
    post:
      tags: [Notifications]
      operationId: markAllNotificationsRead
      security: [{bearerAuth: []}, {cookieAuth: []}]
      responses:
        "204": {description: All notifications marked read.}
  /api/v1/me/notification-preferences:
    get:
      tags: [Notifications]
      operationId: listNotificationPreferences
      security: [{bearerAuth: []}, {cookieAuth: []}]
      responses:
        "200":
          description: Complete notification delivery preference catalog.
          content:
            application/json:
              schema: {type: array, items: {$ref: "#/components/schemas/NotificationPreference"}}
  /api/v1/me/notification-preferences/{type}:
    put:
      tags: [Notifications]
      operationId: updateNotificationPreference
      security: [{bearerAuth: []}, {cookieAuth: []}]
      parameters:
        - {name: type, in: path, required: true, schema: {$ref: "#/components/schemas/NotificationType"}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [in_app_enabled, email_enabled]
              properties:
                in_app_enabled: {type: boolean}
                email_enabled: {type: boolean}
      responses:
        "200": {description: Updated preference., content: {application/json: {schema: {$ref: "#/components/schemas/NotificationPreference"}}}}
        "400": {$ref: "#/components/responses/Error"}
...
  /api/v1/admin/audit-events:
    get:
      tags: [Admin Audit]
      operationId: listAdminAuditEvents
      security:
        - bearerAuth: []
      parameters:
        - name: action
          in: query
          schema:
            $ref: "#/components/schemas/AdminAuditAction"
        - name: actor_user_id
          in: query
          schema:
            type: string
            format: uuid
        - name: target_type
          in: query
          schema:
            $ref: "#/components/schemas/AdminAuditTargetType"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: Admin audit events ordered newest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminAuditEvent"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
...
  /api/v1/tickets/{ticketID}/comments:
    parameters:
      - $ref: "#/components/parameters/TicketID"
    get:
      tags: [Comments]
      operationId: listTicketComments
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: Comments on a ticket.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
        "400":
          $ref: "#/components/responses/Error"
    post:
      tags: [Comments]
      operationId: createComment
      security:
        - bearerAuth: []
      requestBody:
        $ref: "#/components/requestBodies/CreateComment"
      responses:
        "201":
          description: Comment created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          $ref: "#/components/responses/Error"
  /api/v1/comments/{commentID}:
    parameters:
      - $ref: "#/components/parameters/CommentID"
    delete:
      tags: [Comments]
      operationId: deleteComment
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Comment deleted.
        "404":
          $ref: "#/components/responses/Error"
...
  /comments/{commentID}:
    parameters:
      - $ref: "#/components/parameters/CommentID"
    get:
      tags: [ActivityPub]
      operationId: getCommentDocument
      responses:
        "200":
          $ref: "#/components/responses/ActivityPubDocument"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: progo.session
      description: HttpOnly SameSite=Strict browser session cookie.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    metricsBearerAuth:
      type: http
      scheme: bearer
      description: Optional in development; required when METRICS_TOKEN is configured.
...
  parameters:
...
    CommentID:
      name: commentID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    TicketID:
      name: ticketID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  requestBodies:
...
    CreateComment:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [content]
            properties:
              content:
                type: string
...
  responses:
    ActivityPubDocument:
      description: ActivityStreams JSON-LD document.
      content:
        application/activity+json:
          schema:
            $ref: "#/components/schemas/ActivityPubDocument"
        application/ld+json:
          schema:
            $ref: "#/components/schemas/ActivityPubDocument"
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    AdminAuditAction:
      type: string
      enum:
        - user.instance_role_updated
        - federation.domain_blocked
        - federation.domain_unblocked
        - federation.delivery_retried
    AdminAuditEvent:
      type: object
      required: [id, action, target_type, target_id, metadata, created_at]
      properties:
        id:
          type: string
          format: uuid
        actor_user_id:
          type: string
          format: uuid
        action:
          $ref: "#/components/schemas/AdminAuditAction"
        target_type:
          $ref: "#/components/schemas/AdminAuditTargetType"
        target_id:
          type: string
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
    AdminAuditTargetType:
      type: string
      enum: [user, federation_domain, federation_delivery]
    ActivityPubDocument:
      type: object
      additionalProperties: true
    Comment:
      type: object
      required: [id, ap_id, ticket_id, author_id, content, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        ap_id:
          type: string
          format: uri
        ticket_id:
          type: string
          format: uuid
        author_id:
          type: string
          format: uuid
        content:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
...
    NotificationType:
      type: string
      enum:
        - ticket.assigned
        - ticket.status_changed
        - ticket.due_soon
        - ticket.overdue
        - comment.created
        - comment.mentioned
        - project.invited
        - project.role_changed
        - federation.delivery_failed
        - security.event
    Notification:
      type: object
      required: [id, user_id, type, title, body, created_at]
      properties:
        id: {type: string, format: uuid}
        user_id: {type: string, format: uuid}
        actor_id: {type: [string, "null"], format: uuid}
        project_id: {type: [string, "null"], format: uuid}
        ticket_id: {type: [string, "null"], format: uuid}
        type: {$ref: "#/components/schemas/NotificationType"}
        title: {type: string}
        body: {type: string}
        read_at: {type: [string, "null"], format: date-time}
        created_at: {type: string, format: date-time}
    NotificationPreference:
      type: object
      required: [type, in_app_enabled, email_enabled]
      properties:
        type: {$ref: "#/components/schemas/NotificationType"}
        in_app_enabled: {type: boolean}
        email_enabled: {type: boolean}
        updated_at: {type: string, format: date-time}
...